requirements.txt
Django==1.6.7
MySQL-python
django-registration-redux==1.2
South
Pillow
